feat: refactor E2E tests to use relative URLs and simplify page imports
